+++ src/main/java/io/electrum/prepaidutility/api/KeyChangeTokenRequestsResource.java
@@ -16,11 +16,9 @@ import javax.ws.rs.core.UriInfo;
 
 import io.electrum.prepaidutility.model.ErrorDetail;
 import io.electrum.prepaidutility.model.KeyChangeConfirmation;
-import io.electrum.prepaidutility.model.KeyChangeReversal;
 import io.electrum.prepaidutility.model.KeyChangeTokenRequest;
 import io.electrum.prepaidutility.model.KeyChangeTokenResponse;
 import io.electrum.vas.model.BasicAdvice;
-import io.electrum.vas.model.BasicReversal;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
@@ -51,25 +49,12 @@ public abstract class KeyChangeTokenRequestsResource {
       public static final String CONFIRM_KEY_CHANGE = "confirmKeyChange";
       public static final int SUCCESS = 202;
       public static final String RELATIVE_PATH =
-              "/{" + PathParameters.REQUEST_ID +"}/confirmations/{" +PathParameters.CONFIRMATION_ID +"}";
+            "/{" + PathParameters.REQUEST_ID + "}/confirmations/{" + PathParameters.CONFIRMATION_ID + "}";
       public static final String PATH = KeyChangeTokenRequestsResource.PATH + RELATIVE_PATH;
 
       public class PathParameters {
          public static final String REQUEST_ID = "requestId";
          public static final String CONFIRMATION_ID = "confirmationId";
-      }
-   }
-
-   public class ReverseKeyChange {
-      public static final String REVERSE_KEY_CHANGE = "reverseKeyChange";
-      public static final int SUCCESS = 202;
-      public static final String RELATIVE_PATH =
-              "/{" + PathParameters.REQUEST_ID +"}/reversals/{" +PathParameters.REVERSAL_ID +"}";
-      public static final String PATH = KeyChangeTokenRequestsResource.PATH + RELATIVE_PATH;
-
-      public class PathParameters {
-         public static final String REQUEST_ID = "requestId";
-         public static final String REVERSAL_ID = "reversalId";
       }
    }
 
@@ -142,39 +127,4 @@ public abstract class KeyChangeTokenRequestsResource {
             uriInfo);
    }
 
-   @POST
-   @Path(ReverseKeyChange.RELATIVE_PATH)
-   @Consumes({ "application/json" })
-   @Produces({ "application/json" })
-   @ApiOperation(nickname = ReverseKeyChange.REVERSE_KEY_CHANGE, value = "Notifies provider that a key change was not completed successfully.", notes = "Notifies provider that a key change was not completed successfully. This can occur if the original request timed out or if the key change was unsuccessful")
-   @ApiResponses(value = {
-         @ApiResponse(code = ReverseKeyChange.SUCCESS, message = "Accepted", response = BasicReversal.class),
-         @ApiResponse(code = 400, message = "Bad request", response = ErrorDetail.class),
-         @ApiResponse(code = 404, message = "Not found", response = ErrorDetail.class),
-         @ApiResponse(code = 500, message = "Internal Server Error", response = ErrorDetail.class),
-         @ApiResponse(code = 501, message = "Not implemented", response = ErrorDetail.class),
-         @ApiResponse(code = 503, message = "Service Unavailable", response = ErrorDetail.class),
-         @ApiResponse(code = 504, message = "Gateway Timeout", response = ErrorDetail.class) })
-   public void reverseKeyChange(
-         @ApiParam(value = "The randomly generated UUID of this request.", required = true) @PathParam(ReverseKeyChange.PathParameters.REQUEST_ID) String requestId,
-         @ApiParam(value = "The randomly generated UUID of this reversal.", required = true) @PathParam(ReverseKeyChange.PathParameters.REVERSAL_ID) String reversalId,
-         @ApiParam(value = "A key change reversal", required = true) KeyChangeReversal body,
-         @Context SecurityContext securityContext,
-         @Suspended AsyncResponse asyncResponse,
-         @Context Request request,
-         @Context HttpServletRequest httpServletRequest,
-         @Context HttpHeaders httpHeaders,
-         @Context UriInfo uriInfo) {
-
-      getResourceImplementation().reverseKeyChange(
-            requestId,
-            reversalId,
-            body,
-            securityContext,
-            asyncResponse,
-            request,
-            httpServletRequest,
-            httpHeaders,
-            uriInfo);
-   }
 }